Create Project
Wrote simple code for How I want each individual project card to look like.

The idea of this is that the user of the web-app would click the "Download" button on a project of mine that they like, so that it leads them to a new page with more detailed description of project with option of downloading my work.

Full functionality is still not implemented as of yet